backend/requirements.txt
# Core Framework
Flask==3.0.0
Werkzeug==3.0.1
gunicorn==21.2.0

# Security
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
Flask-CORS==4.0.0
Flask-Talisman==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7

# Rate Limiting
Flask-Limiter==3.5.0

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
psycopg2-binary==2.9.9
alembic==1.13.1
SQLAlchemy==2.0.23

# Redis/Caching
redis==5.0.1
Flask-Caching==2.1.0

# ML Dependencies
scikit-learn
numpy<2.0.0
scipy
joblib
pandas

# Document Processing
PyPDF2==3.0.1
python-docx==1.1.0

# Audio/Video Processing
SpeechRecognition==3.10.1
pydub==0.25.1

# Validation
marshmallow==3.20.1
email-validator==2.1.0
python-magic==0.4.27

# Monitoring & Logging
prometheus-flask-exporter==0.23.0
sentry-sdk[flask]==1.39.1
python-json-logger==2.0.7

# API Documentation
flask-swagger-ui==4.11.1
apispec==6.3.1
apispec-webframeworks==1.0.0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
celery==5.3.4
